requirements.txt
aniso8601==9.0.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
flask-restx==1.1.0
gcloud==0.17.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jws==0.1.3
MarkupSafe==2.1.2
oauth2client==3.0.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==3.0.9
Pyrebase4==4.6.0
pyrsistent==0.19.3
python-decouple==3.8
python-jwt==2.0.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
gunicorn